+++ kube/examples/crd_derive.rs
@@ -6,12 +6,16 @@ use serde_derive::{Deserialize, Serialize};
 ///
 /// A struct with our chosen Kind will be created for us, using the following kube attrs
 #[derive(CustomResource, Serialize, Deserialize, Debug, Clone)]
-#[kube(group = "clux.dev", version = "v1", kind = "Foo", namespaced)]
-#[kube(status = "FooStatus")]
-#[kube(scale = r#"{"specReplicasPath":".spec.replicas", "statusReplicasPath":".status.replicas"}"#)]
 #[kube(
+    group = "clux.dev",
+    version = "v1",
+    kind = "Foo",
+    namespaced,
+    status = "FooStatus",
+    scale = r#"{"specReplicasPath":".spec.replicas", "statusReplicasPath":".status.replicas"}"#,
     printcolumn = r#"{"name":"Spec", "type":"string", "description":"name of foo", "jsonPath":".spec.name"}"#
 )]
+#[kube(apiextensions = "v1beta1")] // kubernetes <= 1.16
 pub struct MyFoo {
     name: String,
     info: Option<String>,
@@ -41,7 +45,7 @@ fn verify_crd() {
     use serde_json::{self, json};
     let crd = Foo::crd();
     let output = json!({
-      "apiVersion": "apiextensions.k8s.io/v1",
+      "apiVersion": "apiextensions.k8s.io/v1beta1",
       "kind": "CustomResourceDefinition",
       "metadata": {
         "name": "foos.clux.dev"
@@ -54,22 +58,26 @@ fn verify_crd() {
           "shortNames": [],
           "singular": "foo"
         },
+        "additionalPrinterColumns": [
+          {
+            "description": "name of foo",
+            "JSONPath": ".spec.name",
+            "name": "Spec",
+            "type": "string"
+          }
+        ],
         "scope": "Namespaced",
         "versions": [
           {
-            "additionalPrinterColumns": [
-              {
-                "description": "name of foo",
-                "jsonPath": ".spec.name",
-                "name": "Spec",
-                "type": "string"
-              }
-            ],
             "name": "v1",
             "served": true,
             "storage": true
           }
-        ]
+        ],
+        "subresources": {
+          "scale": {"specReplicasPath":".spec.replicas", "statusReplicasPath":".status.replicas"},
+          "status": {}
+        }
       }
     });
     let outputcrd = serde_json::from_value(output).expect("expected output is valid");