+++ kube/examples/event_informer.rs
@@ -2,9 +2,8 @@
 use k8s_openapi::api::core::v1::Event;
 use kube::{
     api::{ListParams, Resource, WatchEvent},
-    client::APIClient,
-    config,
     runtime::Informer,
+    Client, Configuration,
 };
 
 use futures::{StreamExt, TryStreamExt};
@@ -13,8 +12,7 @@ use futures::{StreamExt, TryStreamExt};
 async fn main() -> anyhow::Result<()> {
     std::env::set_var("RUST_LOG", "info,kube=debug");
     env_logger::init();
-    let config = config::load_kube_config().await?;
-    let client = APIClient::new(config);
+    let client = Client::from(Configuration::infer().await?);
 
     let events = Resource::all::<Event>();
     let lp = ListParams::default();