+++ kube/examples/node_reflector.rs
@@ -2,17 +2,15 @@
 use k8s_openapi::api::core::v1::Node;
 use kube::{
     api::{ListParams, Meta, Resource},
-    client::APIClient,
-    config,
     runtime::Reflector,
+    Client, Configuration,
 };
 
 #[tokio::main]
 async fn main() -> anyhow::Result<()> {
     std::env::set_var("RUST_LOG", "info,kube=debug");
     env_logger::init();
-    let config = config::load_kube_config().await?;
-    let client = APIClient::new(config);
+    let client = Client::from(Configuration::infer().await?);
 
     let resource = Resource::all::<Node>();
     let lp = ListParams::default()