+++ kube/examples/pod_informer.rs
@@ -3,9 +3,8 @@ use futures::{StreamExt, TryStreamExt};
 use k8s_openapi::api::core::v1::Pod;
 use kube::{
     api::{ListParams, Meta, Resource, WatchEvent},
-    client::APIClient,
-    config,
     runtime::Informer,
+    Client, Configuration,
 };
 use std::env;
 
@@ -13,8 +12,7 @@ use std::env;
 async fn main() -> anyhow::Result<()> {
     env::set_var("RUST_LOG", "info,kube=debug");
     env_logger::init();
-    let config = config::load_kube_config().await?;
-    let client = APIClient::new(config);
+    let client = Client::from(Configuration::infer().await?);
     let namespace = env::var("NAMESPACE").unwrap_or("default".into());
 
     let resource = Resource::namespaced::<Pod>(&namespace);