+++ kube/src/lib.rs
@@ -61,6 +61,8 @@ pub type Result<T> = std::result::Result<T, Error>;
 pub mod api;
 pub use api::{Api, Resource};
 pub mod client;
+pub use client::Client;
 pub mod config;
+pub use config::Configuration;
 mod oauth2;
 pub mod runtime;