+++ kube/src/runtime/reflector.rs
@@ -1,11 +1,10 @@
 use crate::{
     api::{ListParams, Meta, ObjectList, Resource, WatchEvent},
-    client::APIClient,
-    Error, Result,
+    Client, Error, Result,
 };
 use futures::{lock::Mutex, StreamExt, TryStreamExt};
-use futures_timer::Delay;
 use serde::de::DeserializeOwned;
+use tokio::time::delay_for;
 
 use std::{collections::BTreeMap, sync::Arc, time::Duration};
 
@@ -41,7 +40,7 @@ where
     K: Clone + DeserializeOwned + Send + Meta,
 {
     state: Arc<Mutex<State<K>>>,
-    client: APIClient,
+    client: Client,
     resource: Resource,
     params: ListParams,
 }
@@ -51,7 +50,7 @@ where
     K: Clone + DeserializeOwned + Meta + Send,
 {
     /// Create a reflector with a kube client on a resource
-    pub fn new(client: APIClient, lp: ListParams, r: Resource) -> Self {
+    pub fn new(client: Client, lp: ListParams, r: Resource) -> Self {
         Reflector {
             client,
             resource: r,
@@ -77,7 +76,7 @@ where
             warn!("Poll error on {}: {}: {:?}", self.resource.kind, e, e);
             // If desynched due to mismatching resourceVersion, retry in a bit
             let dur = Duration::from_secs(10);
-            Delay::new(dur).await;
+            delay_for(dur).await;
             self.reset().await?; // propagate error if this failed..
         }